Log a Warning if a Device Takes Longer than 1 Second to Get Descriptors
The usb.find_devices() function can sometimes take a long time (>30 seconds) to run due to a problematic device taking a long time to read descriptors. Sometimes this can be fixed by unplugging the device and plugging it back in. When enumerating devices, the time taken to read the descriptors is recorded. If a device takes longer than 1 second, a warning is logged which is useful to find the problematic device.